MGS-7531 [#mali] gpu: Integrate Mali DDK r47p0
1. Disallows changing the sharability on the GPU of imported dma-bufs
to BASE_MEM_COHERENT_SYSTEM using KBASE_IOCTL_MEM_FLAGS_CHANGE.
2. Implement full block state support for hardware counters.
3. add imx platform instead of devicetree

+++ drivers/gpu/arm/midgard/arbiter/mali_kbase_arbiter_pm.c
@@ -308,7 +308,7 @@ int kbase_arbiter_pm_early_init(struct kbase_device *kbdev)
 		err = wait_event_timeout(arb_vm_state->vm_state_wait,
 					 arb_vm_state->vm_state ==
 						 KBASE_VM_STATE_INITIALIZING_WITH_GPU,
-					 msecs_to_jiffies(gpu_req_timeout));
+					 msecs_to_jiffies((unsigned int)gpu_req_timeout));
 
 		if (!err) {
 			dev_dbg(kbdev->dev,
