+++ drivers/gpu/arm/midgard/backend/gpu/mali_kbase_js_backend.c
@@ -169,11 +169,11 @@ static enum hrtimer_restart timer_callback(struct hrtimer *timer)
 					 * now. Hard stop the slot.
 					 */
 #if !KBASE_DISABLE_SCHEDULING_HARD_STOPS
-					int ms = js_devdata->scheduling_period_ns / 1000000u;
+					u32 ms = js_devdata->scheduling_period_ns / 1000000u;
 					dev_warn(
 						kbdev->dev,
-						"JS: Job Hard-Stopped (took more than %lu ticks at %lu ms/tick)",
-						(unsigned long)ticks, (unsigned long)ms);
+						"JS: Job Hard-Stopped (took more than %u ticks at %u ms/tick)",
+						ticks, ms);
 					kbase_job_slot_hardstop(atom->kctx, s, atom);
 #endif
 				} else if (ticks == gpu_reset_ticks) {
@@ -204,11 +204,11 @@ static enum hrtimer_restart timer_callback(struct hrtimer *timer)
 					 * ticks. Hard stop the slot.
 					 */
 #if !KBASE_DISABLE_SCHEDULING_HARD_STOPS
-					int ms = js_devdata->scheduling_period_ns / 1000000u;
+					u32 ms = js_devdata->scheduling_period_ns / 1000000u;
 					dev_warn(
 						kbdev->dev,
-						"JS: Job Hard-Stopped (took more than %lu ticks at %lu ms/tick)",
-						(unsigned long)ticks, (unsigned long)ms);
+						"JS: Job Hard-Stopped (took more than %u ticks at %u ms/tick)",
+						ticks, ms);
 					kbase_job_slot_hardstop(atom->kctx, s, atom);
 #endif
 				} else if (ticks == js_devdata->gpu_reset_ticks_dumping) {
