+++ drivers/gpu/arm/midgard/csf/mali_kbase_csf_firmware_log.c
@@ -144,7 +144,7 @@ static ssize_t kbasep_csf_firmware_log_debugfs_read(struct file *file, char __us
 	}
 
 	*ppos += n_read;
-	ret = n_read;
+	ret = (int)n_read;
 
 out:
 	atomic_set(&fw_log->busy, 0);
@@ -178,8 +178,9 @@ static int kbase_csf_firmware_log_mode_write(void *data, u64 val)
 		break;
 	case KBASE_CSF_FIRMWARE_LOG_MODE_AUTO_PRINT:
 	case KBASE_CSF_FIRMWARE_LOG_MODE_AUTO_DISCARD:
-		schedule_delayed_work(&fw_log->poll_work,
-				      msecs_to_jiffies(atomic_read(&fw_log->poll_period_ms)));
+		schedule_delayed_work(
+			&fw_log->poll_work,
+			msecs_to_jiffies((unsigned int)atomic_read(&fw_log->poll_period_ms)));
 		break;
 	default:
 		ret = -EINVAL;
@@ -198,7 +199,7 @@ static int kbase_csf_firmware_log_poll_period_read(void *data, u64 *val)
 	struct kbase_device *kbdev = (struct kbase_device *)data;
 	struct kbase_csf_firmware_log *fw_log = &kbdev->csf.fw_log;
 
-	*val = atomic_read(&fw_log->poll_period_ms);
+	*val = (u64)atomic_read(&fw_log->poll_period_ms);
 	return 0;
 }
 
@@ -263,7 +264,7 @@ static void kbase_csf_firmware_log_poll(struct work_struct *work)
 		return;
 
 	schedule_delayed_work(&fw_log->poll_work,
-			      msecs_to_jiffies(atomic_read(&fw_log->poll_period_ms)));
+			      msecs_to_jiffies((unsigned int)atomic_read(&fw_log->poll_period_ms)));
 }
 
 int kbase_csf_firmware_log_init(struct kbase_device *kbdev)
@@ -382,7 +383,7 @@ void kbase_csf_firmware_log_dump_buffer(struct kbase_device *kbdev)
 		pend = p + read_size;
 		p = buf;
 
-		while (p < pend && (pnewline = memchr(p, '\n', pend - p))) {
+		while (p < pend && (pnewline = memchr(p, '\n', (size_t)(pend - p)))) {
 			/* Null-terminate the string */
 			*pnewline = 0;
